[FRONT] A check was added in the CustomCart.tsx component. Then, a similar check for the presence of data was also added in HomePage.tsx. In cartSlice.ts, a function was added to create a cart for an anonymous user in case they are not registered and don't have a cart at all. After that, this function was called in the OneProductCard.tsx component to create a new cart.

+++ frontend/src/features/Home/HomePage.tsx
@@ -79,7 +79,11 @@ const HomePage = () => {
     const syncProductsToBackend = async () => {
       if (user && products.length > 0 && cart && !synced) {
         setSynced(true);
-        await dispatch(deleteItemsCart({cartId: cart.id, token: user.token})).unwrap();
+
+        if (cart.products.length > 0) {
+          await dispatch(deleteItemsCart({cartId: cart.id, token: user.token})).unwrap();
+        }
+
         for (const product of products) {
           await dispatch(
             addItem({
@@ -95,6 +99,9 @@ const HomePage = () => {
     };
     void syncProductsToBackend();
   }, [dispatch, user, products, cart, synced]);
+
+  console.log(products);
+  console.log(cart);
 
   return (
     <Container>
